+++ lib/attributes/backend/serial/shared.cpp
@@ -9,7 +9,11 @@ using namespace clang;
 
 __attribute__((constructor)) void registerOPENMPSharedHandler() {
     auto ok = registerBackendHandler(
-        TargetBackend::SERIAL, SHARED_ATTR_NAME, serial_subset::handleSharedAttribute);
+        TargetBackend::SERIAL, SHARED_ATTR_NAME, serial_subset::handleSharedDeclAttribute);
+    ok &= registerBackendHandler(
+        TargetBackend::SERIAL, SHARED_ATTR_NAME, serial_subset::handleSharedTypeAttribute);
+    ok &= registerBackendHandler(
+        TargetBackend::SERIAL, SHARED_ATTR_NAME, serial_subset::handleSharedVarAttribute);
 
     // Empty Stmt handler since @shared variable is of attributed type, it is called on DeclRefExpr
     ok &= registerBackendHandler(
