+++ lib/attributes/utils/utils.h
@@ -5,5 +5,6 @@
 #include <clang/AST/AST.h>
 
 namespace oklt {
-const clang::AttributedStmt* getAttributedStmt(SessionStage& s, const clang::Stmt& stmt);
+const clang::AttributedStmt* getAttributedStmt(SessionStage&, const clang::Stmt&);
+std::string getCleanTypeString(clang::QualType);
 }
